+++ nucleoid_built/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.3
+
+- Change MySQL to SQL
+- Features type FieldTypeDart.date
+
 ## 0.2.2
 
 - MySQL model generator features: test table; auto correct table; set and insert operation.